handle option
